JSON as Codable & OSLog
Provide a Codable implementation for JSONObject — a wrapper around a dictionary of String → Any. This makes it possible to define and share structs for the requests between the Pterodactyl library & server.

Using OSLog in the Server so it’s possible to actually see what’s happening in the server during tests.